Set up step velocity system

+++ src/components/step.rs
@@ -8,19 +8,19 @@ pub enum Side {
 }
 
 impl Side {
-    pub fn base_x_component(&self) -> f32 {
+    pub fn base_x_component(&self) -> i32 {
         match self {
-            crate::components::Side::VERTICAL => 0.,
-            crate::components::Side::HORIZONTAL => -1.,
-            crate::components::Side::DIAGONAL => 1.,
+            crate::components::Side::VERTICAL => 0,
+            crate::components::Side::HORIZONTAL => -1,
+            crate::components::Side::DIAGONAL => 1,
         }
     }
 
-    pub fn base_y_component(&self) -> f32 {
+    pub fn base_y_component(&self) -> i32 {
         match self {
-            crate::components::Side::VERTICAL => 1.,
-            crate::components::Side::HORIZONTAL => 0.,
-            crate::components::Side::DIAGONAL => -1.,
+            crate::components::Side::VERTICAL => 1,
+            crate::components::Side::HORIZONTAL => 0,
+            crate::components::Side::DIAGONAL => -1,
         }
     }
 }
